fix: replace RefCell with RwLock to hold AppConfig
In multithreaded (or async) contexts, using RefCell (as in
Arc<RefCell<T>>) might cause panics if multiple borrows are
accessed, since RefCell is meant for single-threaded use,
which is why Clippy was complaining when holding a RefCell
borrow across await points.

The thread-safe alternative to RefCell is either Mutex or RwLock,
while the latter is better in read-heavy contexts, hence this change.